pyproject.toml
[tool.poetry]
name = "bug-report"
version = "0.1.0"
description = ""
authors = ["Andrew Simmons <agsimmons0@gmail.com>"]
packages = [{include = "bug_report", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "~3.2.16"
django-image-cropping = "^1.7"
easy-thumbnails = "^2.8.3"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
